// src/api/openapi/model/orderBundleArticleAddedEvent.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderedArticleSnapshot } from './orderedArticleSnapshot';


export interface OrderBundleArticleAddedEvent { 
    oldOrderId?: number | null;
    orderedArticle: OrderedArticleSnapshot;
    type?: OrderBundleArticleAddedEvent.TypeEnum;
}
export namespace OrderBundleArticleAddedEvent {
    export const TypeEnum = {
        ArticleAdded: 'article_added'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}