// src/api/openapi/model/orderBundleArticleChangedEvent.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderedArticleChange } from './orderedArticleChange';


export interface OrderBundleArticleChangedEvent { 
    name?: string | null;
    'new': OrderedArticleChange;
    old: OrderedArticleChange;
    orderedArticleId: number;
    type?: OrderBundleArticleChangedEvent.TypeEnum;
}
export namespace OrderBundleArticleChangedEvent {
    export const TypeEnum = {
        ArticleChanged: 'article_changed'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}