// src/api/openapi/model/orderBundleArticleRemovedEvent.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderedArticleSnapshot } from './orderedArticleSnapshot';


export interface OrderBundleArticleRemovedEvent { 
    newOrderId?: number | null;
    orderedArticle: OrderedArticleSnapshot;
    type?: OrderBundleArticleRemovedEvent.TypeEnum;
}
export namespace OrderBundleArticleRemovedEvent {
    export const TypeEnum = {
        ArticleRemoved: 'article_removed'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}