// src/api/openapi/model/orderBundleCreatedEvent.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderBundleSource } from './orderBundleSource';


export interface OrderBundleCreatedEvent { 
    deliveryDate: string;
    description: string;
    orderFromId: number;
    orderIds?: Array<number> | null;
    request: boolean;
    source: OrderBundleSource;
    sourceOrderIds?: Array<number> | null;
    type?: OrderBundleCreatedEvent.TypeEnum;
}
export namespace OrderBundleCreatedEvent {
    export const TypeEnum = {
        Created: 'created'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}