// src/api/openapi/model/orderBundleDeliveryDateChangedEvent.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OrderBundleDeliveryDateChangedEvent { 
    'new': string;
    old: string;
    type?: OrderBundleDeliveryDateChangedEvent.TypeEnum;
}
export namespace OrderBundleDeliveryDateChangedEvent {
    export const TypeEnum = {
        DeliveryDateChanged: 'delivery_date_changed'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}