// src/api/openapi/model/orderBundleDescriptionChangedEvent.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OrderBundleDescriptionChangedEvent { 
    'new': string;
    old: string;
    type?: OrderBundleDescriptionChangedEvent.TypeEnum;
}
export namespace OrderBundleDescriptionChangedEvent {
    export const TypeEnum = {
        DescriptionChanged: 'description_changed'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}